requirements.txt
aiohttp
anytree
appdirs
aria2p
js2py
attrdict
beautifulsoup4
cfscrape
feedparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
pyshorteners
python-dotenv
python-magic
python-telegram-bot
gc-python-utils
qbittorrent-api
requests
speedtest-cli
telegraph
tenacity
TgCrypto
torrentool==1.1.0
urllib3
yt_dlp
heroku3
gitpython
